+++ src/pages/my-page/MyPage.tsx
@@ -8,7 +8,7 @@ import { MapPin, Calendar, Settings } from 'lucide-react'
 import { ProductMetaItem } from '@src/components/product/ProductMetaItem'
 import { Tabs } from '@src/components/Tabs'
 import { MY_PAGE_TABS, type MyPageTabId } from '@src/constants/constants'
-import MyPagePanel from './MyPagePanel'
+import MyPagePanel from './components/MyPagePanel'
 import DeleteConfirmModal from '@src/components/modal/DeleteConfirmModal'
 import WithdrawModal, { type WithDrawFormValues } from '@src/components/modal/WithdrawModal'
 import { withDraw } from '@src/api/profile'
